# client/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "david energy",
    "version": "1.0.0",
    "title": "dave-energy"
  },
  "servers": [
    {
      "url": "https://david-energy-staging.herokuapp.com"
    },
    {
      "url": "http://127.0.0.1:5000"
    }
  ],
...
  "paths": {
...
    "/building/{building_id}/annual-energy-demand": {
      "get": {
        "tags": [
          "Building"
        ],
        "summary": "Get building annual-energy-demand",
        "description": "",
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "description": "pass building_id",
            "schema": {
              "type": "integer",
              "example": 5
            }
          },
          {
            "name": "year",
            "in": "query",
            "description": "pass year (int, optional): year=2020 ",
            "schema": {
              "type": "integer",
              "example": "2020"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/annual-energy-demand"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/building/{building_id}/annual-energy-consumption": {
      "get": {
        "tags": [
          "Building"
        ],
        "summary": "Get building annual-energy-consumption",
        "description": "",
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "description": "pass building_id",
            "schema": {
              "type": "integer",
              "example": 5
            }
          },
          {
            "name": "year",
            "in": "query",
            "description": "pass year (int, optional): year=2020 ",
            "schema": {
              "type": "integer",
              "example": "2020"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/annual-energy-consumption"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/building/{building_id}/energy-consumption/compare": {
      "get": {
        "tags": [
          "Building"
        ],
        "summary": "Compare energy consumption for building by year specified in query. NOTE: Log-in required",
        "description": "",
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "description": "pass building_id",
            "schema": {
              "type": "integer",
              "example": 5
            }
          },
          {
            "name": "year1",
            "in": "query",
            "description": "year1=2019",
            "schema": {
              "type": "integer",
              "example": "2019"
            }
          },
          {
            "name": "year2",
            "in": "query",
            "description": "year2=2020 ",
            "schema": {
              "type": "integer",
              "example": "2020"
            }
          }

        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                    "$ref": "#/components/schemas/compare-energy-consumption"
                }
              }
            }
          }
        }
      }
    },
    "/building/{building_id}/energy-demand/compare": {
      "get": {
        "tags": [
          "Building"
        ],
        "summary": "Compare energy demand for building by year specified in query. NOTE: Log-in required",
        "description": "",
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "description": "pass building_id",
            "schema": {
              "type": "integer",
              "example": 5
            }
          },
          {
            "name": "year1",
            "in": "query",
            "description": "year1=2019",
            "schema": {
              "type": "integer",
              "example": "2019"
            }
          },
          {
            "name": "year2",
            "in": "query",
            "description": "year2=2020 ",
            "schema": {
              "type": "integer",
              "example": "2020"
            }
          }

        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                    "$ref": "#/components/schemas/compare-demand-consumption"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "annual-energy-demand": {
          "type": "object",
          "properties": {
            "month_display": {
              "type": "string",
              "example": "J"
            },
            "month_number": {
              "type": "integer",
              "example": 1
            },
            "quantity": {
              "type": "object",
              "properties": {
                "quantity": {
                  "type": "number",
                  "example": 452.0
                },
                "target": {
                  "type": "integer",
                  "example": 400
                },
                "targeted": {
                  "type": "integer",
                  "example": -45
                },
                "ts": {
                  "type": "string",
                  "example": "2020-10-02T09:45:00.000000"
                }
              }
            },
            "unit": {
              "type": "string",
              "example": "kW"
            }
          }
      },
      "annual-energy-consumption": {
          "type": "object",
          "properties": {
            "month_display": {
              "type": "string",
              "example": "J"
            },
            "month_number": {
              "type": "integer",
              "example": 1
            },
            "quantity": {
              "type": "number",
              "example": 192624.0
            },
            "unit": {
              "type": "string",
              "example": "kW"
            }
          }
      },
...
      "compare-energy-consumption": {
          "type": "object",
          "properties": {
            "2019": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "1": {
                    "type": "number",
                    "example": 0.0
                  },
                  "2": {
                    "type": "number",
                    "example": 0.0
                  },
                  "3": {
                    "type": "number",
                    "example": 0.0
                  },
                  "4": {
                    "type": "number",
                    "example": 0.0
                  },
                  "5": {
                    "type": "number",
                    "example": 0.0
                  },
                  "6": {
                    "type": "number",
                    "example": 0.0
                  },
                  "7": {
                    "type": "number",
                    "example": 0.0
                  },
                  "8": {
                    "type": "number",
                    "example": 0.0
                  },
                  "9": {
                    "type": "number",
                    "example": 0.0
                  },
                  "10": {
                    "type": "number",
                    "example": 0.0
                  },
                  "11": {
                    "type": "number",
                    "example": 0.0
                  },
                  "12": {
                    "type": "number",
                    "example": 0.0
                  }
                }
              }
            },
            "2020": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "1": {
                    "type": "number",
                    "example": 1000.0
                  },
                  "2": {
                    "type": "number",
                    "example": 20.0
                  },
                  "3": {
                    "type": "number",
                    "example": 6698.0
                  },
                  "4": {
                    "type": "number",
                    "example": 0.0
                  },
                  "5": {
                    "type": "number",
                    "example": 0.0
                  },
                  "6": {
                    "type": "number",
                    "example": 0.0
                  },
                  "7": {
                    "type": "number",
                    "example": 0.0
                  },
                  "8": {
                    "type": "number",
                    "example": 0.0
                  },
                  "9": {
                    "type": "number",
                    "example": 0.0
                  },
                  "10": {
                    "type": "number",
                    "example": 0.0
                  },
                  "11": {
                    "type": "number",
                    "example": 0.0
                  },
                  "12": {
                    "type": "number",
                    "example": 0.0
                  }
                }
              }
            }
          }
      },
      "compare-demand-consumption": {
          "type": "object",
          "properties": {
            "2019": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "1": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-01-01T05:00:00.000000"
                      }
                    }
                  },
                  "2": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-02-01T05:00:00.000000"
                      }
                    }
                  },
                  "3": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-03-01T05:00:00.000000"
                      }
                    }
                  },
                  "4": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-04-01T05:00:00.000000"
                      }
                    }
                  },
                  "5": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-05-01T05:00:00.000000"
                      }
                    }
                  },
                  "6": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-06-01T05:00:00.000000"
                      }
                    }
                  },
                  "7": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-07-01T05:00:00.000000"
                      }
                    }
                  },
                  "8": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-08-01T05:00:00.000000"
                      }
                    }
                  },
                  "9": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-09-01T05:00:00.000000"
                      }
                    }
                  },
                  "10": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-10-01T05:00:00.000000"
                      }
                    }
                  },
                  "11": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-11-01T05:00:00.000000"
                      }
                    }
                  },
                  "12": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-12-01T05:00:00.000000"
                      }
                    }
                  }
                }
              }
            },
            "2020": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "1": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-01-01T05:00:00.000000"
                      }
                    }
                  },
                  "2": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-02-01T05:00:00.000000"
                      }
                    }
                  },
                  "3": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-03-01T05:00:00.000000"
                      }
                    }
                  },
                  "4": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-04-01T05:00:00.000000"
                      }
                    }
                  },
                  "5": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-05-01T05:00:00.000000"
                      }
                    }
                  },
                  "6": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-06-01T05:00:00.000000"
                      }
                    }
                  },
                  "7": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-07-01T05:00:00.000000"
                      }
                    }
                  },
                  "8": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-08-01T05:00:00.000000"
                      }
                    }
                  },
                  "9": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-09-01T05:00:00.000000"
                      }
                    }
                  },
                  "10": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-10-01T05:00:00.000000"
                      }
                    }
                  },
                  "11": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-11-01T05:00:00.000000"
                      }
                    }
                  },
                  "12": {
                    "type": "object",
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "example": 452.0
                      },
                      "target": {
                        "type": "integer",
                        "example": 400
                      },
                      "targeted": {
                        "type": "integer",
                        "example": -45
                      },
                      "ts": {
                        "type": "string",
                        "example": "2019-12-01T05:00:00.000000"
                      }
                    }
                  }
                }
              }
            }
          }
      },
...
    }
  }
}